refactor: rename file and fix tests
